refactor: moved debounce from hooks to utils
